Add shared preprocessing config module, migrate 19 consumers, fix diffusion build bug
- preprocessing/artifact_paths.py: single source of truth for artifact paths
  and image shape (width/height/channels), default (128,128,16) resolves to
  byte-identical paths as before — zero behavior change for production.
- Migrate preprocessing/gan/classifer/diffusion/evaluation (19 files) off
  locally-duplicated path/shape constants onto the shared config.
- prepare_training_data.py: add --width/--height/--channels CLI, routing
  size-independent artifacts (tsne, gene stats, phenotypes) to the shared
  dir and size-dependent ones to a tagged subdirectory.
- pixel_collision_analysis.py / gene_reconstruction_analysis.py: parameterize
  by target size/config, add exact-vs-averaged collision accounting
  (compute_exact_mask), support single-channel configs for the 256x256x1 /
  512x512x1 comparison requested by advisor.
- diffusion_model.py: fix tf.cast() on a KerasTensor placeholder (invalid
  under Keras 3 Functional API) that was blocking build_unet() entirely;
  switched to keras.ops.cast in the occupancy-mask conditioning paths.
- diffusion_preprocessing_eval.py: modernize off DDPM-era imports removed in
  the EDM2 migration (init_schedule, prepare_batch_ordinal,
  visualize_diffusion_process), fix multichannel display, add
  visualize_noise_journey.
- ARCHITECTURE.md: update to match current EDM2 params (P_mean, sigma_data),
  document occupancy-mask conditioning/enforcement/sampling, rewrite the
  next-steps section into priority tiers.

+++ diffusion/diffusion_model.py
@@ -641,7 +641,7 @@ def build_unet(config):
     h = MPConv2D(channels[0], 3)(x_input)
     # Project binary mask into feature space (use_bias=False so unoccupied positions
     # contribute exactly zero, not a learned offset).
-    h = h + MPConv2D(channels[0], 1, use_bias=False)(tf.cast(mask_input, h.dtype))
+    h = h + MPConv2D(channels[0], 1, use_bias=False)(keras.ops.cast(mask_input, h.dtype))
     
     # Encoder
     skip_connections = []
@@ -717,7 +717,7 @@ def build_unet(config):
     # Hard occupancy enforcement: zero out positions that are structurally empty.
     # Paired with the masked loss in train_step, this guarantees the model never
     # predicts anything at unoccupied channels and is never penalized there.
-    output = output * tf.cast(mask_input, output.dtype)
+    output = output * keras.ops.cast(mask_input, output.dtype)
 
     # Build model — outputs [pred_F, logvar]; logvar is [B,1,1,1], pred_F is [B,H,W,C]
     model = keras.Model(
